+++ particle_tracking_manager/__init__.py
@@ -8,6 +8,7 @@ logging.getLogger('kerchunk').setLevel(logging.WARNING)
 logging.getLogger('opendrift').setLevel(logging.WARNING)
 logging.getLogger('numcodecs').setLevel(logging.WARNING)
 
+from .models.opendrift.config_opendrift import OpenDriftConfig
 from .models.opendrift.opendrift import OpenDriftModel
 # from .the_manager import ParticleTrackingManager
-from .models.opendrift.config_opendrift import OpenDriftConfig
+# import particle_tracking_manager.models.opendrift.config_opendrift