+++ particle_tracking_manager/config_the_manager.py
@@ -49,10 +49,10 @@ class LogLevelEnum(str, Enum):
 # from geojson_pydantic import LineString, Point, Polygon
 
 class TheManagerConfig(BaseModel):
-    model: ModelEnum = Field(ModelEnum.opendrift, description="Lagrangian model software to use for simulation.", ptm_level=1)
-    lon: Optional[float] = Field(-151.0, ge=-180, le=180, description="Central longitude for seeding drifters. Only used if `seed_flag==\"elements\"`.", ptm_level=1, units="degrees_east")
-    lat: Optional[float] = Field(58.0, ge=-90, le=90, description="Central latitude for seeding drifters. Only used if `seed_flag==\"elements\"`.", ptm_level=1, units="degrees_north")
-    geojson: Optional[dict] = Field(None, description="GeoJSON describing a polygon within which to seed drifters. To use this parameter, also have `seed_flag==\"geojson\"`.", ptm_level=1)
+    model: ModelEnum = Field(ModelEnum.opendrift.value, description="Lagrangian model software to use for simulation.", json_schema_extra=dict(ptm_level=1))
+    lon: Optional[float] = Field(-151.0, ge=-180, le=180, description="Central longitude for seeding drifters. Only used if `seed_flag==\"elements\"`.", json_schema_extra=dict(ptm_level=1, units="degrees_east"))
+    lat: Optional[float] = Field(58.0, ge=-90, le=90, description="Central latitude for seeding drifters. Only used if `seed_flag==\"elements\"`.", json_schema_extra=dict(ptm_level=1, units="degrees_north"))
+    geojson: Optional[dict] = Field(None, description="GeoJSON describing a polygon within which to seed drifters. To use this parameter, also have `seed_flag==\"geojson\"`.", json_schema_extra=dict(ptm_level=1))
 #   geojson: Annotated[
 #     Union[Point, LineString, Polygon],
 #     Field(
@@ -60,26 +60,24 @@ class TheManagerConfig(BaseModel):
 #         description="GeoJSON describing a point, line, or polygon for seeding drifters.",  # noqa: E501
 #     ),
     # ]
-    seed_flag: SeedFlagEnum = Field(SeedFlagEnum.elements, description="Method for seeding drifters. Options are \"elements\" or \"geojson\". If \"elements\", seed drifters at or around a single point defined by lon and lat. If \"geojson\", seed drifters within a polygon described by a GeoJSON object.", ptm_level=1)
-    start_time: Optional[datetime] = Field(datetime(2022,1,1), description="Start time for drifter simulation.", ptm_level=1,
-                                           ge=datetime(1999,1,1), le=datetime(2023,1,2))
-    start_time_end: Optional[datetime] = Field(None, description="If used, this creates a range of start times for drifters, starting with `start_time` and ending with `start_time_end`. Drifters will be initialized linearly between the two start times.", ptm_level=2)
-    run_forward: bool = Field(True, description="Run forward in time.", ptm_level=2)
-    time_step: float = Field(5, ge=0.01, le=1440, description="Interval between particles updates, in minutes.", ptm_level=3, units="minutes")
-    time_step_output: float = Field(60, ge=1, le=1440, description="Time step at which element properties are stored and eventually written to file. This must be larger than the calculation time step, and be an integer multiple of this.", ptm_level=3, units="minutes", od_mapping='general:time_step_output_minutes')
-    steps: Optional[int] = Field(None, ge=1, le=10000, description="Maximum number of steps. End of simulation will be start_time + steps * time_step.", ptm_level=1)
-    duration: Optional[str] = Field(None, description="Duration should be input as a string of ISO 8601. The length of the simulation. steps, end_time, or duration must be input by user.", ptm_level=1)
-    end_time: Optional[datetime] = Field(None, description="The end of the simulation. steps, end_time, or duration must be input by user.", ptm_level=1,
-                                           ge=datetime(1999,1,1), le=datetime(2023,1,2))
-    ocean_model: Optional[OceanModelEnum] = Field(OceanModelEnum.CIOFSOP, description="Name of ocean model to use for driving drifter simulation.", ptm_level=1)
-    ocean_model_local: bool = Field(True, description="Set to True to use local version of known `ocean_model` instead of remote version.", ptm_level=3)
-    do3D: bool = Field(False, description="Set to True to run drifters in 3D, by default False for most drift models.", ptm_level=1)
-    use_static_masks: bool = Field(True, description="Set to True to use static masks for known models instead of wetdry masks.", ptm_level=3)
-    output_file: Optional[str] = Field(None, description="Name of file to write output to. If None, default name is used.", ptm_level=3)
-    output_format: OutputFormatEnum = Field(OutputFormatEnum.netcdf, description="Output file format. Options are \"netcdf\" or \"parquet\".", ptm_level=2)
-    use_cache: bool = Field(True, description="Set to True to use cache for storing interpolators.", ptm_level=3)
-    horizontal_diffusivity: Optional[float] = Field(None, description="Horizontal diffusivity for the simulation.", ptm_level=2, od_mapping="drift:horizontal_diffusivity")
-    log_level: LogLevelEnum = Field(LogLevelEnum.INFO, description="Log verbosity", ptm_level=3)
+    seed_flag: SeedFlagEnum = Field(SeedFlagEnum.elements.value, description="Method for seeding drifters. Options are \"elements\" or \"geojson\". If \"elements\", seed drifters at or around a single point defined by lon and lat. If \"geojson\", seed drifters within a polygon described by a GeoJSON object.", json_schema_extra=dict(ptm_level=1))
+    start_time: Optional[datetime] = Field(datetime(2022,1,1), description="Start time for drifter simulation.", json_schema_extra=dict(ptm_level=1))
+    start_time_end: Optional[datetime] = Field(None, description="If used, this creates a range of start times for drifters, starting with `start_time` and ending with `start_time_end`. Drifters will be initialized linearly between the two start times.", json_schema_extra=dict(ptm_level=2))
+    run_forward: bool = Field(True, description="Run forward in time.", json_schema_extra=dict(ptm_level=2))
+    time_step: float = Field(5, ge=0.01, le=1440, description="Interval between particles updates, in minutes.", json_schema_extra=dict(ptm_level=3, units="minutes"))
+    time_step_output: float = Field(60, ge=1, le=1440, description="Time step at which element properties are stored and eventually written to file. This must be larger than the calculation time step, and be an integer multiple of this.", json_schema_extra=dict(ptm_level=3, units="minutes"))
+    steps: Optional[int] = Field(None, ge=1, le=10000, description="Maximum number of steps. End of simulation will be start_time + steps * time_step.", json_schema_extra=dict(ptm_level=1))
+    duration: Optional[str] = Field(None, description="Duration should be input as a string of ISO 8601. The length of the simulation. steps, end_time, or duration must be input by user.", json_schema_extra=dict(ptm_level=1))
+    end_time: Optional[datetime] = Field(None, description="The end of the simulation. steps, end_time, or duration must be input by user.", json_schema_extra=dict(ptm_level=1))
+    ocean_model: Optional[OceanModelEnum] = Field(OceanModelEnum.CIOFSOP.value, description="Name of ocean model to use for driving drifter simulation.", json_schema_extra=dict(ptm_level=1))
+    ocean_model_local: bool = Field(True, description="Set to True to use local version of known `ocean_model` instead of remote version.", json_schema_extra=dict(ptm_level=3))
+    do3D: bool = Field(False, description="Set to True to run drifters in 3D, by default False for most drift models.", json_schema_extra=dict(ptm_level=1))
+    use_static_masks: bool = Field(True, description="Set to True to use static masks for known models instead of wetdry masks.", json_schema_extra=dict(ptm_level=3))
+    output_file: Optional[str] = Field(None, description="Name of file to write output to. If None, default name is used.", json_schema_extra=dict(ptm_level=3))
+    output_format: OutputFormatEnum = Field(OutputFormatEnum.netcdf.value, description="Output file format. Options are \"netcdf\" or \"parquet\".", json_schema_extra=dict(ptm_level=2))
+    use_cache: bool = Field(True, description="Set to True to use cache for storing interpolators.", json_schema_extra=dict(ptm_level=3))
+    horizontal_diffusivity: Optional[float] = Field(None, description="Horizontal diffusivity for the simulation.", json_schema_extra=dict(ptm_level=2))
+    log_level: LogLevelEnum = Field(LogLevelEnum.INFO.value, description="Log verbosity", json_schema_extra=dict(ptm_level=3))
     # TODO: change log_level to "verbose" or similar
 
     
@@ -89,14 +87,14 @@ class TheManagerConfig(BaseModel):
         title="Horizontal Diffusivity",
         ge=0,
         le=100000,
-        units="m2/s",
+        json_schema_extra=dict(units="m2/s"),
     )
     
     stokes_drift: bool = Field(
         default=True,
         description="Advection elements with Stokes drift (wave orbital motion).",
         title="Stokes Drift",
-        ptm_level=2, 
+        json_schema_extra=dict(ptm_level=2), 
     )
 
     z: Optional[float] = Field(
@@ -105,8 +103,9 @@ class TheManagerConfig(BaseModel):
         title="Z",
         le=0,
         ge=-10000,
+        json_schema_extra=dict(
         units="m",
-        ptm_level=1
+        ptm_level=1)
     )
 
     number: int = Field(
@@ -114,15 +113,17 @@ class TheManagerConfig(BaseModel):
         description="The number of elements for the simulation.",
         title="Number",
         ge=1,
+        json_schema_extra=dict(
         units=1,
         ptm_level=1, 
-        od_mapping="seed:number",
+        )
     )
 
-    class Config:
-        validate_defaults = True
-        use_enum_values=True
-        extra="forbid"
+    model_config = {
+        "validate_defaults": True,
+        "use_enum_values": True,
+        "extra": "forbid",
+    }
 
     @model_validator(mode='after')
     def check_config_seed_flag_elements(self) -> Self:
@@ -222,17 +223,24 @@ class TheManagerConfig(BaseModel):
     
     @computed_field
     def ocean_model_simulation(self) -> OceanModelSimulation:
+        # import pdb; pdb.set_trace()
         inputs = {
             # "ocean_model_config": self.ocean_model_config,
-            "oceanmodel_lon0_360": self.ocean_model_config.oceanmodel_lon0_360,
+            # "oceanmodel_lon0_360": self.ocean_model_config.oceanmodel_lon0_360,
             "lon": self.lon,
             "lat": self.lat,
             "start_time": self.start_time,
             "end_time": self.end_time,
             "ocean_model_local": self.ocean_model_local,
         }
-        return ocean_model_simulation_mapper.get(self.ocean_model)(**inputs)    
-        # return ocean_model_simulation_mapper[self.ocean_model](**inputs)    
+        # return ocean_model_simulation_mapper.get(self.ocean_model)(**inputs)    
+        return ocean_model_simulation_mapper[self.ocean_model](**inputs)
+    
+    @model_validator(mode='after')
+    def select_ocean_model_simulation_on_init(self) -> Self:
+        """Select ocean model simulation based on ocean_model."""
+        self.ocean_model_simulation
+        return self
     
     @model_validator(mode='after')
     def assign_horizontal_diffusivity(self) -> Self: